first version of layout
