+++ src/DataService/CRMLiteAppService/Controllers/ProfileController.cs
@@ -12,37 +12,38 @@ using Microsoft.EntityFrameworkCore;
 
 namespace CRMLiteAppService.Controllers
 {
-	[Authorize(Policy = "ApiUser")]
-	[Route("api/[controller]/[action]")]
-	public class ProfileController : Controller
-	{
-		private readonly ClaimsPrincipal _caller;
-		private readonly ApplicationDbContext _appDbContext;
+    [Authorize(Policy = "ApiUser")]
+    [Route("api/[controller]/[action]")]
+    public class ProfileController : Controller
+    {
+        private readonly ClaimsPrincipal _caller;
+        private readonly ApplicationDbContext _appDbContext;
 
-		public ProfileController(UserManager<AppUser> userManager, ApplicationDbContext appDbContext, IHttpContextAccessor httpContextAccessor)
-		{
-			_caller = httpContextAccessor.HttpContext.User;
-			_appDbContext = appDbContext;
-		}
+        public ProfileController(UserManager<AppUser> userManager, ApplicationDbContext appDbContext, IHttpContextAccessor httpContextAccessor)
+        {
+            _caller = httpContextAccessor.HttpContext.User;
+            _appDbContext = appDbContext;
+        }
 
-		// GET api/profile/me
-		[HttpGet]
-		public async Task<IActionResult> Me()
-		{
-			// retrieve the user info
-			var userId = _caller.Claims.Single(c => c.Type == "id");
-			var customer = await _appDbContext.Customers.Include(c => c.Identity).SingleAsync(c => c.Identity.Id == userId.Value);
+        // GET api/profile/me
+        [HttpGet]
+        public async Task<IActionResult> Me()
+        {
+            // retrieve the user info
+            var userId = _caller.Claims.Single(c => c.Type == "id");
+            var customer = await _appDbContext.Customers.Include(c => c.Identity).SingleAsync(c => c.Identity.Id == userId.Value);
 
-			return new OkObjectResult(new
-			{
-				customer.Identity.FirstName,
-				customer.Identity.LastName,
-				customer.Identity.PictureUrl,
-				customer.Identity.FacebookId,
-				customer.Location,
-				customer.Locale,
-				customer.Gender
-			});
-		}
-	}
+            return new OkObjectResult(new
+            {
+                Message = "This is secure API and user data!",
+                customer.Identity.FirstName,
+                customer.Identity.LastName,
+                customer.Identity.PictureUrl,
+                customer.Identity.FacebookId,
+                customer.Location,
+                customer.Locale,
+                customer.Gender
+            });
+        }
+    }
 }
