+++ src/WebApp/src/layouts/Empty.vue
@@ -1,0 +1,16 @@
+<template>
+  <v-app>
+    <v-content>
+      <v-container fluid>
+        <router-view></router-view>
+      </v-container>
+    </v-content>
+  </v-app>
+</template>
+
+<script lang="ts">
+import { Component, Vue } from 'vue-property-decorator';
+
+@Component
+export default class LayoutEmpty extends Vue {}
+</script>
