+++ src/WebApp/src/router.ts
@@ -1,13 +1,25 @@
 import Vue from 'vue';
 import Router from 'vue-router';
-import Home from './pages/Home.vue';
 import store from './store/store';
+
+import Login from './pages/account/Login.vue';
+import Register from './pages/account/Register.vue';
+import Home from './pages/Home.vue';
+import NotFound from './pages/NotFound.vue';
 
 Vue.use(Router);
 
 const router = new Router({
   mode: 'history',
   routes: [
+    {
+      path: '*',
+      component: NotFound,
+      beforeEnter: (to: any, from: any, next: any) => {
+        store.commit('setLayout', 'layout-empty');
+        next();
+      },
+    },
     {
       path: '/',
       name: 'home',
@@ -17,16 +29,18 @@ const router = new Router({
     {
       path: '/register',
       name: 'registrationForm',
+      component: Register,
       beforeEnter: (to: any, from: any, next: any) => {
-        store.commit('setLayout', 'layout-register');
+        store.commit('setLayout', 'layout-empty');
         next();
       },
     },
     {
       path: '/login',
       name: 'loginForm',
+      component: Login,
       beforeEnter: (to: any, from: any, next: any) => {
-        store.commit('setLayout', 'layout-login');
+        store.commit('setLayout', 'layout-empty');
         next();
       },
     },
@@ -38,10 +52,11 @@ router.beforeEach((to: any, from: any, next: any) => {
   if (to.matched.some((record: any) => record.meta.requiresAuth)) {
     // this route requires auth, check if logged in
     // if not, redirect to login page.
-    if (!store.getters['auth/isAuthenticated']) {
+    if (!store.getters['auth/isLoggedIn']) {
       next({
         path: '/login',
         query: { redirect: to.fullPath },
+        params: { redirect: to.fullPath },
       });
     }
   }
